Use cursor style only on CodeMirror editor
This commit removes the global use of `document.style.cursor`
and instead provides a Facet to set the cursor style
via the content attributes of the editor.

This ensures that other parts of the page/document won't have their cursor style changed due to the editor and is more sound.